Move decomp definitions for GP_SERV_CHAR_PC
